Consolidate components into single package
Move all components from subdirectories into the main components/
package so users only need one import. All components now use
`package components` and live directly in the components/ directory.
