+++ components/form_field.go
@@ -1,4 +1,4 @@
-package form_field
+package components
 
 import "github.com/rohanthewiz/element"
