+++ components/table.go
@@ -1,4 +1,4 @@
-package table
+package components
 
 import (
 	"fmt"
